# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FedaPay API",
        "description": "Description de mon API",
        "version": "1.0.0"
    },
    "paths": {
        "fedapay/api/transaction": {
            "post": {
                "tags": [
                    "FedaPay"
                ],
                "summary": "Créer une transaction",
                "description": "Crée une transaction avec les données reçues du formulaire",
                "operationId": "createTransaction",
                "requestBody": {
                    "description": "Données de la transaction",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Transaction for john.doe@example.com"
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": 2000
                                    },
                                    "currency": {
                                        "required": [
                                            "iso"
                                        ],
                                        "properties": {
                                            "iso": {
                                                "type": "string",
                                                "example": "XOF"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "callback_url": {
                                        "type": "string",
                                        "example": "https://www.monsite.com/callback"
                                    },
                                    "customer": {
                                        "required": [
                                            "email"
                                        ],
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "john.doe@example.com"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirection vers le lien de paiement"
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/fedapay/api/callback": {
            "get": {
                "tags": [
                    "FedaPay"
                ],
                "summary": "Gérer le retour d'appel de transaction",
                "description": "Traite le retour d'appel de la transaction et effectue les actions appropriées",
                "operationId": "handleCallback",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID de la transaction",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Statut de la transaction",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "approved",
                                "canceled"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Paiement effectué"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "302": {
                        "description": "Redirection vers une page appropriée"
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transaction/status/{status}": {
            "get": {
                "tags": [
                    "FedaPay"
                ],
                "summary": "Transaction Status",
                "description": "Displays the transaction status page based on the returned status",
                "operationId": "getTransactionStatus",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Transaction status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "approved"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}